Separate types from mongoose schemas
