+++ demos/ios-webview/procaptcha/procaptcha.html
@@ -8,7 +8,7 @@
 <form action="" method="POST">
 <input type="text" name="email" placeholder="Email" />
 <input type="password" name="password" placeholder="Password" />
-<div class="procaptcha" data-sitekey="5FNFBC97JQoJu7LBapycXbT66N9xzQH2tyFv1yNvBxqa8tQR" data-captcha-type="image"></div>
+<div class="procaptcha" data-sitekey="5GCP69mhanZqJqnTvaaGvJCJZWSahz6xE2c7bTGETqSB5KDK" data-captcha-type="image"></div>
 <br />
 <input type="submit" value="Submit" />
 </form>
