+++ packages/provider/src/tasks/frictionless/frictionlessTasksUtils.ts
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import type { ScoreComponents } from "@prosopo/types-database";
+import type { ScoreComponents } from "@prosopo/types";
 
 export const computeFrictionlessScore = (
 	scoreComponents:
