+++ packages/provider/src/tests/unit/tasks/decisionMachine/decisionMachineRunner.unit.test.ts
@@ -14,15 +14,13 @@
 
 import {
 	CaptchaType,
+	type DecisionMachineArtifact,
 	type DecisionMachineCaptchaType,
 	DecisionMachineDecision,
 	DecisionMachineRuntime,
 	DecisionMachineScope,
 } from "@prosopo/types";
-import type {
-	DecisionMachineArtifact,
-	IProviderDatabase,
-} from "@prosopo/types-database";
+import type { IProviderDatabase } from "@prosopo/types-database";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { DecisionMachineRunner } from "../../../../tasks/decisionMachine/decisionMachineRunner.js";
