+++ packages/types/src/provider/index.ts
@@ -15,3 +15,4 @@ export * from "./accounts.js";
 export * from "./api.js";
 export * from "./scheduler.js";
 export * from "./detection.js";
+export * from "./database.js";
